# [Simulation] A better heuristic for permuting gates before the simple DP